+++ tests/integrations/inputs/input-select.spec.ts
@@ -75,7 +75,6 @@ describe("Input Select Testing", () => {
 		})
 		await app.vm.$nextTick()
 		await wait(233)
-		console.log(app.html())
 		expect(app.text()).toBe(defaultOptions[0].label);
 	})
 	test("Открытие списка выборки по нажатию на иконку", async () => {
@@ -88,12 +87,12 @@ describe("Input Select Testing", () => {
 		expect(app.text()).toBe(defaultOptions.map(a => a.label).join(''))
 	})
 	test("Открытие списка выборки по нажатию space при фокусе на поле", async () => {
-		await app.get('.input-select').trigger('keyup.space');
+		await app.get('.input-select').trigger('keydown.space');
 		expect(app.text()).toBe(defaultOptions.map(a => a.label).join(''))
 	})
 	test("Переключение значений при нажатии клавиши вниз при фокусе на поле.", async () => {
 		function triggerDown() {
-			return app.get('.input-select').trigger('keydown.down');
+			return app.get('.input-select').trigger('keydown', {key: "ArrowDown"});
 		}
 
 		await triggerDown()
@@ -144,10 +143,9 @@ describe("Input Select Testing", () => {
 	})
 
 	test("Disabled select", async () => {
-
 		form.disable()
 		await app.vm.$nextTick()
-		expect(app.find('.input-select_disabled').exists()).toBe(true)
+		expect(app.find('.vf-input-select_disabled').exists()).toBe(true)
 
 		function triggerDown() {
 			return app.get('.input-select').trigger('keydown.down');
@@ -160,7 +158,7 @@ describe("Input Select Testing", () => {
 	test("Error Select", async () => {
 		expect(form.validate()).toBe(false);
 		await app.vm.$nextTick()
-		expect(app.find('.input-select_error').exists()).toBe(true)
+		expect(app.find('.vf-input-select_error').exists()).toBe(true)
 		expect(app.text()).toBe(STORE.requiredMessage)
 	})
 
@@ -186,14 +184,14 @@ describe("Input Select Testing", () => {
 		expect(input.element.getAttribute('tabindex')).toBe("0")
 	})
 	test("Press down should change value", async () => {
-		await input.trigger('keydown.down');
+		await input.trigger('keydown', {key: "ArrowDown"});
 		expect(form.getValueByName(name)).toBe(defaultOptions[0].value)
 
-		await input.trigger('keydown.down');
+		await input.trigger('keydown', {key: "ArrowDown"});
 		expect(form.getValueByName(name)).toBe(defaultOptions[1].value)
 	})
 	test("Press up should change value", async () => {
-		await input.trigger('keydown.up');
+		await input.trigger('keydown', {key: "ArrowUp"});
 		expect(form.getValueByName(name)).toBe(defaultOptions[2].value)
 	})
 	test("Press up and down not work in disabled status", async () => {
@@ -279,5 +277,109 @@ describe("Input Select Testing", () => {
 		}, Promise.resolve())
 
 		expect(form.getValueByName(name)).toEqual(defaultOptions.map(item => item.value).slice(0, 2))
+	})
+
+	test("Shift + ArrowDown: _ -> 0 = _, 1", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+			limit: 2
+		}))
+		const form = (wrap.vm as any).form as Form;
+		currentItem = wrap.find('.container-input-select-current')
+		await currentItem.trigger('keydown.shift.down');
+
+		expect(form.getValueByName(name)).toEqual([defaultOptions[0].value]);
+	})
+	test("Shift + ArrowDown: 1 -> 0 = 1, 1", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+		}))
+		const form = (wrap.vm as any).form as Form;
+		form.setValues({
+			[name]: [defaultOptions[1].value]
+		})
+		await app.vm.$nextTick()
+		await wrap.find('.container-input-select-current').trigger('click');
+		currentItem = wrap.find('.container-input-select-current')
+		await currentItem.trigger('keydown.shift', {
+			key: "ArrowDown"
+		});
+
+		expect(form.getValueByName(name)).toEqual([defaultOptions[1].value, defaultOptions[2].value]);
+	})
+	test("Shift + ArrowDown: 0 -> 1 = 0, 0", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+		}))
+		const form = (wrap.vm as any).form as Form;
+		form.setValues({
+			[name]: [defaultOptions[1].value]
+		})
+		await app.vm.$nextTick()
+		await wrap.find('.container-input-select-current').trigger('click');
+		currentItem = wrap.find('.container-input-select-current')
+
+		await currentItem.trigger('keydown.ctrl', { key: "ArrowUp" })
+		await currentItem.trigger('keydown.shift', {
+			key: "ArrowDown"
+		});
+
+		expect(form.getValueByName(name)).toEqual([]);
+	})
+	test("Shift + ArrowDown: 0 -> 0 = 1, 0", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+		}))
+		const form = (wrap.vm as any).form as Form;
+		await wrap.find('.container-input-select-current').trigger('click');
+		currentItem = wrap.find('.container-input-select-current')
+
+		await currentItem.trigger('keydown.ctrl', { key: "ArrowDown" })
+		await currentItem.trigger('keydown.shift', {
+			key: "ArrowDown"
+		});
+		expect(form.getValueByName(name)).toEqual([defaultOptions[0].value]);
+		await currentItem.trigger('keydown.shift', {
+			key: "ArrowDown"
+		});
+		expect(form.getValueByName(name)).toEqual([defaultOptions[0].value, defaultOptions[1].value]);
+	})
+	test("Shift + ArrowDown: 1 -> 1 = 0, 1", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+		}))
+		const form = (wrap.vm as any).form as Form;
+		await wrap.find('.container-input-select-current').trigger('click');
+		currentItem = wrap.find('.container-input-select-current')
+
+		await currentItem.trigger('keydown.shift', { key: "ArrowDown" });
+		await currentItem.trigger('keydown.shift', { key: "ArrowDown" });
+
+		await currentItem.trigger('keydown.ctrl', { key: "ArrowUp" });
+
+		expect(form.getValueByName(name)).toEqual([defaultOptions[0].value, defaultOptions[1].value]);
+		await currentItem.trigger('keydown.shift', { key: "ArrowDown" });
+		expect(form.getValueByName(name)).toEqual([defaultOptions[1].value]);
+	})
+
+	test("Double SPACE press should not make effect for multiple", async () => {
+		const wrap = defaultMount(defineSelectComponent({
+			multiple: true,
+			options: defaultOptions,
+			limit: 2
+		}))
+		const form = (wrap.vm as any).form as Form;
+		form.setValues({
+			[name]: []
+		})
+		currentItem = wrap.find('.container-input-select-current')
+		await currentItem.trigger('keydown.space');
+		await currentItem.trigger('keydown.space');
+		expect(form.getValueByName(name).length).toBe(0)
 	})
 })